Refactor workflow generator
