+++ buildstockbatch/test/test_inputs/enforce-validate-measures-good-2-with-anchors.yml
@@ -46,7 +46,7 @@ workflow_generator:
       include_timeseries_weather: true
 
     reporting_measures:
-      - measure_dir_name: QOIReport
+      - measure_dir_name: TimeseriesCSVExport
 
     server_directory_cleanup:
       retain_in_idf: false
