+++ buildstockbatch/workflow_generator/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from .commercial import CommercialDefaultWorkflowGenerator  # noqa F041
-from .residential_hpxml import ResidentialHpxmlWorkflowGenerator  # noqa F041
+from .commercial.commercial import CommercialDefaultWorkflowGenerator  # noqa F041
+from .residential.residential_hpxml import ResidentialHpxmlWorkflowGenerator  # noqa F041
