+++ buildstockbatch/workflow_generator/residential/residential_hpxml_schema.yml
@@ -1,0 +1,91 @@
+build_existing_model: include('build-existing-model-spec', required=False)
+emissions: list(include('emission-scenario-spec'), required=False)
+utility_bills: list(include('utility-bill-scenario-spec'), required=False)
+measures: list(include('measure-spec'), required=False)
+reporting_measures: list(include('measure-spec'), required=False)
+simulation_output_report: include('sim-output-report-spec', required=False)
+server_directory_cleanup: include('server-dir-cleanup-spec', required=False)
+debug: bool(required=False)
+---
+build-existing-model-spec:
+    simulation_control_timestep: enum(60, 30, 20, 15, 12, 10, 6, 5, 4, 3, 2, 1, required=False)
+    simulation_control_run_period_begin_month: int(required=False)
+    simulation_control_run_period_begin_day_of_month: int(required=False)
+    simulation_control_run_period_end_month: int(required=False)
+    simulation_control_run_period_end_day_of_month: int(required=False)
+    simulation_control_run_period_calendar_year: int(required=False)
+    add_component_loads: bool(required=False)
+emission-scenario-spec:
+    scenario_name: str(required=True)
+    type: str(required=True)
+    elec_folder: str(required=True)
+    gas_value: num(required=False)
+    propane_value: num(required=False)
+    oil_value: num(required=False)
+    wood_value: num(required=False)
+utility-bill-scenario-spec:
+    scenario_name: str(required=True)
+    simple_filepath: str(required=False)
+    detailed_filepath: str(required=False)
+    elec_fixed_charge: num(required=False)
+    elec_marginal_rate: num(required=False)
+    gas_fixed_charge: num(required=False)
+    gas_marginal_rate: num(required=False)
+    propane_fixed_charge: num(required=False)
+    propane_marginal_rate: num(required=False)
+    oil_fixed_charge: num(required=False)
+    oil_marginal_rate: num(required=False)
+    wood_fixed_charge: num(required=False)
+    wood_marginal_rate: num(required=False)
+    pv_compensation_type: enum('NetMetering', 'FeedInTariff', required=False)
+    pv_net_metering_annual_excess_sellback_rate_type: enum('User-Specified', 'Retail Electricity Cost', required=False)
+    pv_net_metering_annual_excess_sellback_rate: num(required=False)
+    pv_feed_in_tariff_rate: num(required=False)
+    pv_monthly_grid_connection_fee_units: enum('$', '$/kW', required=False)
+    pv_monthly_grid_connection_fee: num(required=False)
+sim-output-report-spec:
+    timeseries_frequency: enum('none', 'timestep', 'hourly', 'daily', 'monthly', required=False)
+    include_timeseries_total_consumptions: bool(required=False)
+    include_timeseries_fuel_consumptions: bool(required=False)
+    include_timeseries_end_use_consumptions: bool(required=False)
+    include_timeseries_emissions: bool(required=False)
+    include_timeseries_emission_fuels: bool(required=False)
+    include_timeseries_emission_end_uses: bool(required=False)
+    include_timeseries_hot_water_uses: bool(required=False)
+    include_timeseries_total_loads: bool(required=False)
+    include_timeseries_component_loads: bool(required=False)
+    include_timeseries_unmet_hours: bool(required=False)
+    include_timeseries_zone_temperatures: bool(required=False)
+    include_timeseries_airflows: bool(required=False)
+    include_timeseries_weather: bool(required=False)
+    include_timeseries_resilience: bool(required=False)
+    timeseries_timestamp_convention: enum('start', 'end', required=False)
+    timeseries_num_decimal_places: int(required=False)
+    add_timeseries_dst_column: bool(required=False)
+    add_timeseries_utc_column: bool(required=False)
+    output_variables: list(include('output-var-spec'), required=False)
+output-var-spec:
+    name: str(required=True)
+measure-spec:
+    measure_dir_name: str(required=True)
+    arguments: map(required=False)
+server-dir-cleanup-spec:
+    retain_in_osm: bool(required=False)
+    retain_in_idf: bool(required=False)
+    retain_pre_process_idf: bool(required=False)
+    retain_eplusout_audit: bool(required=False)
+    retain_eplusout_bnd: bool(required=False)
+    retain_eplusout_eio: bool(required=False)
+    retain_eplusout_end: bool(required=False)
+    retain_eplusout_err: bool(required=False)
+    retain_eplusout_eso: bool(required=False)
+    retain_eplusout_mdd: bool(required=False)
+    retain_eplusout_mtd: bool(required=False)
+    retain_eplusout_rdd: bool(required=False)
+    retain_eplusout_shd: bool(required=False)
+    retain_eplusout_msgpack: bool(required=False)
+    retain_eplustbl_htm: bool(required=False)
+    retain_stdout_energyplus: bool(required=False)
+    retain_stdout_expandobject: bool(required=False)
+    retain_schedules_csv: bool(required=False)
+    debug: bool(required=False)
